Test: keep one recording interceptor of each kind, not eight
The generator test project had three byte-identical RecordingWriteInterceptor
declarations, three RecordingReadInterceptor and two RecordingMethodInterceptor.
Two of each predate this branch; this branch added another while describing it
as shared, which it was not, because the existing copies kept their own.

All eight collapse into one file. No behaviour change: every copy was
identical, so the tests assert on exactly what they did before.
