+++ src/Namotion.Interceptor.Generator.Tests/ExplicitInterfaceBehaviorTests.cs
@@ -327,26 +327,4 @@ public class ExplicitInterfaceBehaviorTests
         Assert.Equal(Rank.Senior, rankProperty.GetValue());
     }
 
-    private sealed class RecordingReadInterceptor : IReadInterceptor
-    {
-        public List<(string PropertyName, object? Value)> Reads { get; } = [];
-
-        public TProperty ReadProperty<TProperty>(ref PropertyReadContext<TProperty> context, ReadInterceptionDelegate<TProperty> next)
-        {
-            var value = next(ref context);
-            Reads.Add((context.Property.Name, value));
-            return value;
-        }
-    }
-
-    private sealed class RecordingWriteInterceptor : IWriteInterceptor
-    {
-        public List<(string PropertyName, object? Value)> Writes { get; } = [];
-
-        public void WriteProperty<TProperty>(ref PropertyWriteContext<TProperty> context, WriteInterceptionDelegate<TProperty> next)
-        {
-            Writes.Add((context.Property.Name, context.NewValue));
-            next(ref context);
-        }
-    }
 }
